+++ onesec-bridge/src/bridge/icp-to-evm/validateReceiptStep.ts
@@ -3,7 +3,7 @@ import * as fromCandid from "../../fromCandid";
 import { type _SERVICE as OneSec } from "../../generated/candid/onesec/onesec.did";
 import type { Details, EvmChain, StepStatus } from "../../types";
 import { BaseStep, err, ICP_CALL_DURATION_MS, ok, sleep } from "../shared";
-import { TransferStep } from "./transfer-step";
+import { TransferStep } from "./transferStep";
 
 export class ValidateReceiptStep extends BaseStep {
   private delayMs: number = 1_000;
